+++ libs/components/lookup/src/lib/modules/lookup/lookup-show-more-async-modal.component.ts
@@ -6,6 +6,7 @@ import {
   OnInit,
   TemplateRef,
 } from '@angular/core';
+import { SkyLogService } from '@skyux/core';
 import { SkyModalInstance } from '@skyux/modals';
 
 import { Subject, Subscription } from 'rxjs';
@@ -58,17 +59,27 @@ export class SkyLookupShowMoreAsyncModalComponent implements OnInit, OnDestroy {
 
   #ngUnsubscribe = new Subject<void>();
 
+  #logSvc: SkyLogService;
+
   #offset = 0;
 
   constructor(
     public modalInstance: SkyModalInstance,
     public context: SkyLookupShowMoreNativePickerAsyncContext,
-    changeDetector: ChangeDetectorRef
+    changeDetector: ChangeDetectorRef,
+    logSvc: SkyLogService
   ) {
     this.#changeDetector = changeDetector;
+    this.#logSvc = logSvc;
   }
 
   public ngOnInit(): void {
+    if (this.context.idProperty === undefined) {
+      this.#logSvc.error(
+        "The lookup component's 'idProperty' input is required when `enableShowMore` and 'searchAsync' are used together."
+      );
+    }
+
     this.repeaterItemTemplate = this.context.userConfig.itemTemplate || null;
     this.searchText = this.context.initialSearch;