+++ libs/components/modals/src/lib/modules/modal/modal-host-context.ts
@@ -1,0 +1,11 @@
+import { Injectable } from '@angular/core';
+
+/**
+ * Provided by the modal service to give the modal host
+ * component additional context and features.
+ * @internal
+ */
+@Injectable()
+export class SkyModalHostContext {
+  constructor(public teardownCallback: () => void) {}
+}